Ok.. time for the next major installment of the Abdera2 rework.
This batch of changes introduces two major new dependencies and a whole slew of related changes. Primarily, the two new dependencies are the Google Guava Library (com.google.common.*) and the Joda-Time library (org.joda.time.*). Joda-Time replaces the old DateTime class that was there. This is a major change that affects the backwards compatibility of the FOM APIs. The reason this was done is simple: Joda-Time is a whole hell of a lot better than the implementation that was there and offers significant functionality beyond anything we could have implemented. When Abdera first emerged, there were no first-rate ISO 8601 DateTime implementations available in open source, now there are. 

The other new dependency (Guava) provides for significantly more powerful applications than what we had. I'll go into detail on this in a separate email. Expect the tight integration with Guava to continue.

Further, there is a significant amount of attention given to integrating Guava and Joda-Time in very interesting ways. Again, I'll go into this in a separate email.

Another significant change that has been made is an emphasis on the use of Builder classes to generate Immutable Thread-safe instances of commonly used objects. Many existing objects that previously were Mutable (e.g. the extension geo.* objects) have been refactored into immutable objects constructed with a Builder pattern. The point behind this change is to make the code much more reliable, readable, and... well... just better overall.